Suggested Delegation from `Elm_Close()`
This makes the status more meaningful, and its hierarchical nature is consistent with `Fld_Close` in 930475ba9f09a0c0a2dfd3947a6220335467945e.

We do make assumptions that the internal state is consistent, as with `.Kind` and so forth; rather than explicitly using `Fld_Close()` on `e.Field` and then `Elm_Close()` on `e`.  But if the algorithm is properly implemented, the state will always be accurate.