Deactivated `Elm_Close()` As Function
In the end, virtually all conditions should be determined by the parsing block where this is invoked; as opposed to having two logics in parallel that might discoordinate.  Furthermore, we want the operations of each block to be as transparent as possible.

But in the future, if the element's ".Kind" controls the parsing flow, then coordination is ensured.  Then `Elm_Close()` can be a function again.