Suggested `Fld_Spec()` to Interpret Enumerated Fields
When the user specifies something like "{1,vbSunday,dddd the dst of mmmm}" the result would be something like "Sunday the 1st of June".  In other words, this syntax permits enumerations from VBA as arguments, which are passed to (say) the Format() function.  There are three ways to specify such an argument:
* Enumeration: The (unqualified) VBA code for the enumerated value.
* Name: The full name of the value, as used conversationally.
* Abbreviation: An abbreviation of the full name in 3 characters or fewer.

Perhaps we should also permit qualifications like "FormatMode.fmtVbFormat".