Noted Native Defaults for Calendar Specifiers
We must likely determine some other way to represent a missing argument altogether.  Now 0 does work for the custom specifiers ("FormatMode" and "PositionKind") because their 0s are reserved for uninitialized ("_Unknown").  But it fails for the native specifiers ("VbDayOfWeek" and "VbFirstWeekOfYear") because their 0s are already used for the explicit "vbUseSystem*" option, which is itself a valid specification.