Suggested `Expr_Cap()`
This automates a complex process of checking whether an expression is encapsulated.  We must often clean it to be certain, since whitespace interferes with the boundaries.  But if this succeeds, then any defused syntax must be updated to reflect the cleaning; and if it fails, then the effects of cleaning must be undone.