Unsuggested Cloning Arguments in `Fld_Close()` Loop
The `Fld_Close*()` delegate for each argument should properly `*Close()` that argument and leave it "cleaned up".  Thus it is (likely) better to modify the original array in that process, than to preserve it with raw expressions.

The only situation where this could backfire is if, somehow, a later argument needed to reference the raw data of an earlier argument, while the earlier was still raw and before it was cleaned up.