Synchronized (De)activation of Nested Defusal
Defusal should only occur within the very outermost braces; while any nested braces are faithfully rendered as characters, and need only be balanced.

While this was originally the case, the logic was not fully synchronized: defusal was activated at *every* opening brace, and only deactivated at the *outermost* closing brace.  As a result, we would add "dfuNest" to "dfu" redundantly, and "dfu" would eventually reach an ambiguous value: the next power of 2.